created sub-package signer and observer for each chain
