+++ zetaclient/chains/bitcoin/fee.go
@@ -30,9 +30,11 @@ const (
 	bytesPerWitness         = 108 // each additional witness incurs about 108 bytes and it may vary
 	defaultDepositorFeeRate = 20  // 20 sat/byte is the default depositor fee rate
 
-	outTxBytesMin = uint64(239)  // 239vB == EstimateSegWitTxSize(2, 2, toP2WPKH)
-	outTxBytesMax = uint64(1543) // 1543v == EstimateSegWitTxSize(21, 2, toP2TR)
-	outTxBytesAvg = uint64(245)  // 245vB is a suggested gas limit for zeta core
+	OutTxBytesMin = uint64(239)  // 239vB == EstimateSegWitTxSize(2, 2, toP2WPKH)
+	OutTxBytesMax = uint64(1543) // 1543v == EstimateSegWitTxSize(21, 2, toP2TR)
+	OutTxBytesAvg = uint64(245)  // 245vB is a suggested gas limit for zeta core
+
+	DynamicDepositorFeeHeight = 834500 // DynamicDepositorFeeHeight contains the starting height (Bitcoin mainnet) from which dynamic depositor fee will take effect
 )
 
 var (
