--- zetaclient/chains/bitcoin/observer_db_test.go
@@ -1,76 +1,0 @@
-package bitcoin
-
-import (
-	"strconv"
-	"testing"
-
-	"github.com/btcsuite/btcd/btcjson"
-	"github.com/stretchr/testify/suite"
-	clienttypes "github.com/zeta-chain/zetacore/zetaclient/types"
-	"gorm.io/driver/sqlite"
-	"gorm.io/gorm"
-)
-
-const tempSQLiteDbPath = "file::memory:?cache=shared"
-const numOfEntries = 2
-
-type BitcoinObserverDBTestSuite struct {
-	suite.Suite
-	db          *gorm.DB
-	submittedTx map[string]btcjson.GetTransactionResult
-}
-
-func TestBitcoinObserverDB(t *testing.T) {
-	suite.Run(t, new(BitcoinObserverDBTestSuite))
-}
-
-func (suite *BitcoinObserverDBTestSuite) SetupTest() {
-	suite.submittedTx = map[string]btcjson.GetTransactionResult{}
-
-	db, err := gorm.Open(sqlite.Open(tempSQLiteDbPath), &gorm.Config{})
-	suite.NoError(err)
-
-	suite.db = db
-
-	err = db.AutoMigrate(&clienttypes.TransactionResultSQLType{})
-	suite.NoError(err)
-
-	//Create some Transaction entries in the DB
-	for i := 0; i < numOfEntries; i++ {
-		txResult := btcjson.GetTransactionResult{
-			Amount:          float64(i),
-			Fee:             0,
-			Confirmations:   0,
-			BlockHash:       "",
-			BlockIndex:      0,
-			BlockTime:       0,
-			TxID:            strconv.Itoa(i),
-			WalletConflicts: nil,
-			Time:            0,
-			TimeReceived:    0,
-			Details:         nil,
-			Hex:             "",
-		}
-		r, _ := clienttypes.ToTransactionResultSQLType(txResult, strconv.Itoa(i))
-		dbc := suite.db.Create(&r)
-		suite.NoError(dbc.Error)
-		suite.submittedTx[strconv.Itoa(i)] = txResult
-	}
-}
-
-func (suite *BitcoinObserverDBTestSuite) TearDownSuite() {}
-
-func (suite *BitcoinObserverDBTestSuite) TestSubmittedTx() {
-	var submittedTransactions []clienttypes.TransactionResultSQLType
-	err := suite.db.Find(&submittedTransactions).Error
-	suite.NoError(err)
-
-	for _, txResult := range submittedTransactions {
-		r, err := clienttypes.FromTransactionResultSQLType(txResult)
-		suite.NoError(err)
-		want := suite.submittedTx[txResult.Key]
-		have := r
-
-		suite.Equal(want, have)
-	}
-}
