+++ zetaclient/chains/evm/observer/observer_db_test.go
@@ -1,4 +1,4 @@
-package evm
+package observer
 
 import (
 	"strconv"
